Move All Definition to Separate Files
This commit moves all the definitions from Definitions.Swift to separate files. The files associated with StrongHold have the Strong Prefix, the generic definitions do not have prefixes.

Issue: #68
