+++ PowerScout/Models/Definitions/StrongScoreLocationType.swift
@@ -1,0 +1,19 @@
+//
+//  StrongScoreLocationType.swift
+//  PowerScout
+//
+//  Created by Srinivas Dhanwada on 3/2/18.
+//  Copyright © 2018 FRC Team 525. All rights reserved.
+//
+
+import Foundation
+
+enum ScoreLocation: Int {
+    case unknown = 0, batter, courtyard, defenses
+    
+    func toString() -> String {
+        return (self == .batter)    ? "Batter"    :
+            (self == .courtyard) ? "Courtyard" :
+            (self == .defenses)  ? "Defenses"  : "Unknown"
+    }
+}
