+++ PowerScout/Models/Definitions/UpdateType.swift
@@ -1,0 +1,13 @@
+//
+//  UpdateType.swift
+//  PowerScout
+//
+//  Created by Srinivas Dhanwada on 3/2/18.
+//  Copyright © 2018 FRC Team 525. All rights reserved.
+//
+
+import Foundation
+
+enum UpdateType : Int {
+    case none = 0, teamInfo, fieldSetup, finalStats, actionsEdited, autonomous, teleop
+}
